Replace the existing PRNG (Xorshift31) with (minimal) PCG-32.

+++ modules/visual_script/visual_script_builtin_funcs.cpp
@@ -802,14 +802,14 @@ void VisualScriptBuiltinFunc::exec_func(BuiltinFunc p_func,const Variant** p_inp
 		case VisualScriptBuiltinFunc::MATH_SEED: {
 
 			VALIDATE_ARG_NUM(0);
-			uint32_t seed=*p_inputs[0];
+			uint64_t seed=*p_inputs[0];
 			Math::seed(seed);
 
 		} break;
 		case VisualScriptBuiltinFunc::MATH_RANDSEED: {
 
 			VALIDATE_ARG_NUM(0);
-			uint32_t seed=*p_inputs[0];
+			uint64_t seed=*p_inputs[0];
 			int ret = Math::rand_from_seed(&seed);
 			Array reta;
 			reta.push_back(ret);
